refactor: move search button back into SearchBox instead of being its own component since SearchIconButton would not have reusable in multiple locations
