--- app/components/ui/molecules/SearchIconButton.md
@@ -1,5 +1,0 @@
-Search button for use within the People Picker
-
-```js
-<SearchIconButton />
-```
